Reservation Creation Data Nullification Workaround
It's Messy but it works! This wasn't my feature anyways.
I copied the dto to a new version of itself on new page, to avoid the null issue. The one lingering issue with this I couldn't figure out was the loss of the CustomerId when returning to home(logged in) when the reservation was complete, leading to errors when trying to create a second reservation or view reservations: solution - logout on reservation completion!